refactor(config): split config_manager.py into focused modules (#567)
config_manager.py had grown to ~1300 lines, ~3x the project's 400-line
guideline. Split it into cohesive modules while keeping all public import
paths stable:

- models.py: Pydantic config models (LLM/MCP/RAG/tool-approval/file-extractor)
  plus the resolve_env_var() value helper
- settings.py: AppSettings pydantic-settings model
- config_loader.py: ConfigManager with lazy, cached file loaders
- config_manager.py: thin entry point re-exporting every public symbol plus
  the module-level singleton and getter functions

Dependency direction is loader -> settings -> models (no cycles). AppSettings
text and ConfigManager logic are unchanged, so behavior is identical and
existing config tests pass without modification. All four modules are now
under 500 lines.
